Implementacao de autenticacao e primeiros testes com HATEOAS

+++ backend/resources/friends.py
@@ -11,12 +11,6 @@ class Friends(Resource):
         required=True,
         help="Friend's name is required."
     )
-    # parser.add_argument(
-    #     'user_id',
-    #     type=int,
-    #     required=True,
-    #     help='User id is required.'
-    # )
 
     @jwt_required
     def post(self):
@@ -33,12 +27,21 @@ class Friends(Resource):
             return {'message': 'The application could not save this friend.'}, 500
         return friend.describe(), 201
     
+    @jwt_required
     def get(self):
-        return {'friends': [friend.describe() for friend in FriendModel.find_all()]}
+        user_id = get_jwt_identity()
+        return {'friends': [friend.describe() for friend in FriendModel.find_all(user_id=user_id)]}
 
 class SingleFriend(Resource):
+    @jwt_required
     def delete(self, id):
         friend = FriendModel.find_by_id(id)
         if friend:
-            friend.delete_from_db()
-        return {'message': 'Friend deleted successfully.'}
+            if friend.user_id == get_jwt_identity():
+                try:
+                    friend.delete_from_db()
+                except:
+                    return {'message': 'This friend cannot be deleted.'}, 403
+                return {'message': 'Friend deleted successfully.'}, 204 # No Content
+            return {'message': 'This user cannot delete this friend.'}, 403 # Forbidden
+        return {'message': 'There is no frind with the given id.'}, 400 # Bab Request
